Test async file downloads
The complex behaviours within a file are all tested with the simpler
sync method, but most actual requests go through async_file, so it's
probably a good idea to prove it works.

In passing, make additional requests for an already-fully-downloaded
file behave sensibly (though in general we shouldn't be trying to
double-download anyway).