Use GitHub API to check mirror freshness when available
If people have git references to a number of public GitHub repositories,
this turns some multi-round-trip git communications into single HTTP
requests.

If the repository isn't public, this wastes an HTTP request because we
still need to continue and do the mirror update -- but it remembers the
failure and doesn't try again, so it's only a one-off cost.