Build gems inside a locked environment containing their dependencies
This is particularly important to ensure we override/ignore any Gemfile
the gem might've shipped while we try to run rake -- beyond referencing
gems we might not have, it almost certainly contains a `gemspec` call,
which is likely to go poorly.

It matters most when we're finding the right rake gem while building a
Rakefile, but it seems like a good idea in general.