Give `lock` control over which packages get updated
We now respect the existing entries in the lockfile, and only update
when we're told to update everything by a certain level (--major etc),
or a specific packages is listed on the command line (or where necessary
to satisfy other packages).

Note these rules are conceptually similar to Bundler's, but [currently?]
intentionally different in detail.