Prepare a pre-alpha release to test the pipeline
Move development dependencies to the Gemfile: we're not packaging the
Rakefile or the tests, so listing their dependencies in the gemspec
seems pointlessly noisy.

Also trim the `files` list. This is risky because it increases the
chances a future release will omit an important file, but it avoids
carrying a variety of gunk, so seems worthwhile. (Immediately
interesting because it skips bin/, which contains a symlink... but
skipping the gemspec and Gemfile is nice too.)

Bumped to 2.0.0.alpha0, because 1.* was taken by the previous 'gel' gem.