Properly format errors from rubygems binstubs
Ideally people will arrive via our binstub, but if they e.g. run a
command supplied by a gem installed in rubygems but not in gel, the
rubygems binstub will hit our activate_bin_path, which will then fail to
find the relevant gem.

The fact it fails is correct, but we don't want to messily leak our
internal exception in the process.