Keep libs in stable order while shuffling versions
A gem might have two instances of the same name, in different paths: if
so, we need to make sure we keep choosing the right one.

My currently-known example is psych: it has both a psych.rb and a
psych.$(DLEXT), and having `require "psych"` load the wrong one is a
recipe for a bad time.