Don't expect Gem constant to be present while loading Gel
First, we don't use rubygems/deprecate, so don't require it -- that just
moves the error to a different file, though.

So we also defer the require of gel/catalog, because it (indirectly)
loads marshal_hacks, which expects Gem to be there. This one should
ultimately be fixed by replacing the marshal-hacks mechanism with a
custom marshal parser. But this rearrangement will make it happy for
now.

Fixes #61