flock the file while we're modifying it
Because we're seeking and appending, things can get confused if a second
process is attempting to update the same file at the same time. This
could mean we end up waiting for another process to finish what it's
doing, but that's fine: it's doing the very thing we'd otherwise be
spending our cycles on, so the wait isn't wasted.