Provide enough empty shells that rake-compiler can (partly) work
We're currently still not aiming to be able to produce a packaged gem --
but we can fake out the parts rake-compiler needs while it's just trying
to set up its rake tasks.

The goal is to allow a rake-compiler-using gem (such as Puma) to work in
local use, and support `rake compile`, under Gel. Actual packaging for
distribution is best left to Bundler & Rubygems.

Fixes #77