Automatically update Gemfile.lock when out of date
When a gem is added or removed from the Gemfile, the Gemfile.lock now updates automatically.
This does mean that gel will now write the Gemfile.lock more often which means more dependency solving.
This behavior more closely matches bundler and prevents confusion around when your Gemfile and Gemfile.lock might not be in sync.