Pad out a bit more API compatibility
For chruby, change Gem.default_dir -- there's still not really a good
answer to that question (because we don't have the same directory
structure as Rubygems), so for now we can just continue to balance the
least-worst interpretation based on how known callers are using it.

For yard and solargraph, some more general API expansion:
Gem::Specification.find_by_path is a reasonable thing to want. Being
able to eval a gemspec is less reasonable, but we can still do our best
to semi-support it.