Specifically handle case-insensitivity during extraction
Separate it from our general "we're unexpectedly trying to overwrite an
existing file!" panic, and realise what's wrong.

Show the user the two conflicting filenames. It's much easier to
immediately understand the problem that way.

Only complain if the file seems important: we're not the gem police.
If it's going to break our ability to require something, we still choose
to hard-fail, because otherwise we might be blamed for weird behaviour
later. But if it appears to just be a supporting file, then it's better
that we continue and hope for the best: it might be a valuable
runtime-loaded data file, but chances are good it's just a test or
something, and that's why no-one has complained to the author / why it
hasn't been fixed.