Maybe fix a false circular-dependency warning
Occasionally a run of the test suite will show a circular require
warning, despite the mentioned file not appearing again in the reported
backtrace. Presumably that means it's happening on a different thread...
and that seems like it at least theoretically could happen in the real
world too. Autoloads should make it safe, regardless.