Rewrite stub handling again
TIL shebang only passes a single argument on Linux, spaces included --
so there's no portable way to go through `env` and also include
parameters.

That's okay: we'll just allow 'gel <path-to-binstub>' as a valid command
format, extract the stub name, and then invoke Stub directly. This does
mean a backwards-incompatible change to the stub format: for now I've
just broken compatibility, but we _could_ do something messier to keep
both working.

We can keep the recent 'ruby -S' fix, with a twist: that doesn't need to
be on the first line, so the operative "ruby" is now on the second.

While we're in the area, allow Gel's binstubs to work with 'ruby -S'
when Gel isn't automatically loaded. In addition to preventing an
edge-case error, this completes the journey of no longer exposing lib/
on the Rubygems-managed load path.