Respect required_ruby_version on installed gems
This is important when there are multiple platform matches, and the
"best" platform match is excluded on this version but is valid (and
potentially has been installed) on other Ruby versions present on this
machine.

Specifically, this came up on a rails/rails checkout that has nokogiri
locked to 1.12.5. 1.12.5-x64_64-darwin is my best platform match, and
works great for Ruby 3.0, but after switching to 3.1 things got sad.