Use the self-alias trick to allow warning-free redefinition
It's highly improbable that we'd be racing with some other code while
loading, but still seems technically better to not leave Kernel.require
missing for a moment.

Also alias both copies of our newly-added method, so they too can be
replaced without generating warnings. I'm not sure whether it's
deliberate, but I think rubygems currently ends up allowing this, so
let's just go with it.