Never auto-resolve while activating an executable
This can make commands suddenly take a surprisingly long time to
execute, even when the executable in question is not part of the lock.

If we're reasonably confident the intended command is not locked, we can
run it despite an outdated lock; otherwise, we should fail and ask the
user to update.