Fix subprocess hanging
I believe the hanging is caused by an issue I learned about in Rust's
Stdio:

> Writing more than a pipe buffer’s worth of input to stdin without also
> reading stdout and stderr at the same time may cause a deadlock. This
> is an issue when running any program that doesn’t guarantee that it
> reads its entire stdin before writing more than a pipe buffer’s worth
> of output. The size of a pipe buffer varies on different targets.

Since the previous code would not do any reads on the pipe until _after_
waiting for the child to finish, we can deadlock because the child is
waiting on a read to be made because it has written as much as it can.

The new implementation uses `detach` to wait for the child in a
background thread, and as long as that thread is alive it continues to
read from the pipe. Once the thread ends (because the child process
exited), we do one more read to make sure the child didn't write
anything in between the last `alive?` returning true and it returning
false.